// src/core/types/BankRedirectionUrls.ts
export default interface BankRedirectionUrls {
  authorisationUrl: string;
  deepLinkAuthorisationUrl?: string | null;
  deepLinkAuthorisationUrlIOS?: string | null;
  businessAppDeepLinkAuthorisationUrl?: string | null;
  businessAppDeepLinkAndroidAuthorisationUrl?: string | null;
  businessAppDeepLinkAuthorisationUrlIOS?: string | null;
}